Check optional dimensionaloty reduction method parameters

Checks if optional parameters passed to dimensionality reduction method are valid and returns a list of all default and optinally set parameters.

checkParams(params, method)

Arguments

params
list of parameters to check; list names must be identical to argument names of chosen dimensionality reduction method.

method
Dimensionality reduction method character to be applied; one of DiffusionMap, DRR, ICA, LLE, Isomap, LaplacianEigenmap, MDS, PCA, kPCA, nMDS, tSNE and UMAP.

Value

named list with default and specified parameters taken by specified dimensionality reduction method

Details

checkParams checks if specified parameters are valid parameters of the underlying dimensionality reduction methods. A list of the methods is provided below:

Diffusion Map: diffuse

Dimensionality reduction by regression (DRR): drr

Independent component analysis (ICA): fastICA

Local liner embedding (LLE): lle

Isomap: isomap

Laplacian Eigenmap: spec.emb and make.kNNG

Multi-dimensional scaling (MDS) : cmdscale

Principal component analysis (PCA): prcomp

Kernel PCA (kPCA): kpca

non-metrix MDS (nMDS): metaMDS

t- stochastic neighbourhood embedding (tSNE): Rtsne

Uniform manifold approximation and projection (umap): umap

Examples

params_DRR <- checkParams(list(cv.folds=10), method="DRR")
params_ICA <- checkParams(list(fun="logcosh", maxit=500), method="ICA")