Estimate stability

Takes the output of subsetDimReduction and finds the stable low-dimension components across dimensionality reduced subsets of the original data.

estimateStability(dr, threshold, procrustes = FALSE, verbose = FALSE)

Arguments

dr
[list] either output of subsetDimReduction or manually generated list with dimensionality reduction results on different subsets of the same dataset. Each list entry is a matrix with row and column names; row.names are crucial to allow for the comparison of the lower dimensional embedding across common sample sets.

threshold
[double] Threshold for stability between 0 and 1 or vector of thresholds between 0 and 1.

procrustes
[logical] indication whether a procrustes transformation to match the lower dimensional representations of common samples between subsets should be performed.

verbose
[logical] If set, progress messages are printed to standard out

Value

named list with [1] stability: list with one dataframe per threshold, where each dataframe contains the stability estimate of the low-dimensional components for that threshold value, [2] corr: data.frame with correlation of the low-dimensional components for all pairwise comparisons and components.