Wrapper function for dimensionality reduction methods

Wrapper function for dimensionality reduction methods

methodsDimReduction(Y, ndim, distY = dist(Y, method = dist.method),
  dist.method = "euclidean", method = c("DiffusionMap", "DRR", "ICA",
  "LLE", "Isomap", "LaplacianEigenmap", "MDS", "PCA", "kPCA", "nMDS",
  "tSNE", "UMAP"), optN = NULL, verbose = FALSE, params = NULL)

Arguments

Y
\[N x P\] data matrix for which the dimensionality of P should be reduced

ndim
maximum dimensionality [integer] to retain in the data; large values can cause long computation times.

distY
[dist] object of class dist containing pairwise distances of Y used for methods DiffusionMap, Isomap, MDS and nMDS; if non specified, stats::dist with Euclidean distance applied to supplied Y.

dist.method
[character] method for computing the distance matrix; one of euclidean, maximum, manhattan, canberra, binary or minkowski; see dist for details.

method
Dimensionality reduction method [character] to be applied; one of DiffusionsMaps, DRR, ICA, LLE, Isomap, LaplacianEigenmap, MDS, PCA, kPCA, nMDS and tSNE.

optN
optimal number [integer] of neighbours to consider for dimensionality reduction; relevant for methods LLE, LaplacianEigenmaps, Isomap and tSNE.

verbose
[logical] If set, progress messages are printed to standard out.

params
[list] optional additional parameters for dimensionality reduction methods; see details.

Value

named list with dimensionality reduced phenotypes (reducedY) and object returned by specified dimensionality reduction method (results) with additional output, see details.

Details

methodsDimReduction wraps around the following implementations of the dimensionality reduction methods it provides: * Diffusion Map: diffuse * Dimensionality reduction by regression (DRR): drr * Independent component analysis (ICA): fastICA * Local liner embedding (LLE): lle * Isomap: isomap * Laplacian Eigenmap: spec.emb and make.kNNG * Multi-dimensional scaling (MDS) : cmdscale * Principal component analysis (PCA): prcomp * Kernel PCA (kPCA): kpca * non-metrix MDS (nMDS): metaMDS * t- stochastic neighbourhood embedding (tSNE): Rtsne * Uniform manifold approximation and projection (umap): umap